* development:
  fix: allow 0-conf in blockchain db (#3680)
  test: fix cucumber WalletQuery.feature (#3677)
  test: fix `wait for` step (#3673)
  perf(comms)!: optimise connection establishment (#3658)
  fix: remove noise negotiation for debugging on bad wire mode (#3657)
  feat: add follow on checkpoint (#3676)
  ci: fix tari collectibles build (#3670)
  feat: example to generate vanity node_ids (#3654)
  fix: prefer configured seeds over dns seeds (#3662)
  docs: update RFC-0230_HTLC to use TariScript (#3622)
  feat: add invoke write method to proxy (#3667)
  ci: add tauri build (#3669)
  test: fix cucumber metadata signature (#3665)
  feat: add asset proxy (#3659)
  test: fix cucumber (#3660)

+++ comms/src/protocol/network_info.rs
@@ -25,10 +25,10 @@
 pub struct NodeNetworkInfo {
     /// Major protocol version. This indicates the protocol version that is supported by this node. A peer MAY reject
     /// the connection if a remote peer advertises a different major version number.
-    pub major_version: u32,
+    pub major_version: u8,
     /// Minor protocol version. A version number that represents backward-compatible protocol changes. A peer SHOULD
     /// NOT reject the connection if a remote peer advertises a different minor version number.
-    pub minor_version: u32,
+    pub minor_version: u8,
     /// The byte that MUST be sent (outbound connections) or MUST be received (inbound connections) for a connection to
     /// be established. This byte cannot be 0x46 (E) because that is reserved for liveness.
     /// Default: 0x00
